+++ frontend/src/lib/components/apps/editor/settingsPanel/ComponentPanel.svelte
@@ -29,6 +29,9 @@
 	import { Delete, ExternalLink } from 'lucide-svelte'
 	import GridCondition from './GridCondition.svelte'
 	import { isTriggerable } from './script/utils'
+	import { inferDeps } from '../appUtilsInfer'
+	import EvalV2InputEditor from './inputEditor/EvalV2InputEditor.svelte'
+	import type { ResultAppInput } from '../../inputType'
 
 	export let componentSettings: { item: GridItem; parent: string | undefined } | undefined =
 		undefined
@@ -94,7 +97,9 @@
 	let viewCssOptions = false
 
 	$: extraLib =
-		componentSettings?.item?.data?.componentInput?.type === 'template' && $worldStore
+		(componentSettings?.item?.data?.componentInput?.type === 'template' ||
+			componentSettings?.item?.data?.componentInput?.type === 'templatev2') &&
+		$worldStore
 			? buildExtraLib(
 					$worldStore?.outputsById ?? {},
 					componentSettings?.item?.data?.id,
@@ -124,6 +129,35 @@
 	const hasInteraction = componentSettings?.item.data.type
 		? isTriggerable(componentSettings?.item.data.type)
 		: false
+
+	let evalV2editor: EvalV2InputEditor | undefined = undefined
+
+	function transformToFrontend() {
+		if (componentSettings?.item.data.componentInput) {
+			const id = componentSettings?.item?.data?.id
+			let appInput: ResultAppInput = {
+				...componentSettings.item.data.componentInput,
+				type: 'runnable',
+				runnable: {
+					type: 'runnableByName',
+					name: `Eval of ${id}`,
+					inlineScript: {
+						path: `${id}_eval`,
+						content: `return ${componentSettings?.item.data.componentInput?.['expr']}`,
+						language: 'frontend',
+						refreshOn: componentSettings?.item.data.componentInput?.['connections']?.map((c) => {
+							return {
+								id: c.componentId,
+								key: c.id
+							}
+						})
+					}
+				},
+				fields: {}
+			}
+			componentSettings.item.data.componentInput = appInput
+		}
+	}
 </script>
 
 <svelte:window on:keydown={keydown} />
@@ -168,30 +202,64 @@
 
 				{#if componentSettings.item.data.componentInput}
 					<ComponentInputTypeEditor
+						{evalV2editor}
 						bind:componentInput={componentSettings.item.data.componentInput}
 					/>
 
-					<div class="flex flex-col w-full gap-2 my-2">
+					<div class="flex flex-col w-full gap-2 mt-2">
 						{#if componentSettings.item.data.componentInput.type === 'static'}
 							<StaticInputEditor
 								fieldType={componentInput?.fieldType}
 								subFieldType={componentInput?.subFieldType}
 								format={componentInput?.format}
 								bind:componentInput={componentSettings.item.data.componentInput}
-								noVariablePicker
 							/>
-						{:else if componentSettings.item.data.componentInput.type === 'template' && componentSettings.item.data.componentInput !== undefined}
+						{:else if componentSettings.item.data?.componentInput?.type === 'template' || componentSettings.item.data?.componentInput?.type === 'templatev2'}
 							<div class="py-1 min-h-[28px] rounded border border-1 border-gray-500">
 								<TemplateEditor
 									fontSize={12}
 									bind:code={componentSettings.item.data.componentInput.eval}
 									{extraLib}
+									on:change={(e) => {
+										if (componentSettings?.item.data.componentInput?.type === 'templatev2') {
+											inferDeps(
+												'`' + e.detail.code + '`',
+												$worldStore.outputsById,
+												componentSettings?.item.data.componentInput,
+												app
+											)
+										}
+									}}
 								/>
 							</div>
+							{#if componentSettings.item.data?.componentInput?.type === 'templatev2'}
+								{#if componentSettings.item.data?.componentInput.connections?.length > 0}
+									<div class="flex flex-wrap gap-2 items-center">
+										<div class="text-2xs text-tertiary">Re-evaluated on changes to:</div>
+										<div class="flex flex-wrap gap-1">
+											{#each componentSettings.item.data?.componentInput.connections as connection (connection.componentId + '-' + connection.id)}
+												<span
+													class="inline-flex items-center rounded-md px-2 py-0.5 text-xs font-medium border"
+													>{connection.componentId + '.' + connection.id}</span
+												>
+											{/each}
+										</div>
+									</div>
+								{/if}
+							{/if}
 						{:else if componentSettings.item.data.componentInput.type === 'connected' && component.componentInput !== undefined}
 							<ConnectedInputEditor
 								bind:componentInput={componentSettings.item.data.componentInput}
 							/>
+						{:else if componentSettings.item.data.componentInput.type === 'evalv2' && component.componentInput !== undefined}
+							<EvalV2InputEditor
+								bind:this={evalV2editor}
+								id={component.id}
+								bind:componentInput={componentSettings.item.data.componentInput}
+							/>
+							<a class="text-2xs" on:click={transformToFrontend} href="#"
+								>transform to a frontend script</a
+							>
 						{:else if componentSettings.item.data.componentInput?.type === 'runnable' && component.componentInput !== undefined}
 							<RunnableInputEditor
 								appComponent={component}
@@ -205,28 +273,24 @@
 				{#key $stateId}
 					{#if componentSettings.item.data.componentInput?.type === 'runnable'}
 						{#if Object.keys(componentSettings.item.data.componentInput.fields ?? {}).length > 0}
-							<div class="border w-full">
-								<PanelSection
-									title={`Runnable Inputs (${
-										Object.keys(componentSettings.item.data.componentInput.fields ?? {}).length
-									})`}
-								>
-									<svelte:fragment slot="action">
-										<Tooltip>
-											The runnable inputs are inferred from the inputs of the flow or script
-											parameters this component is attached to.
-										</Tooltip>
-									</svelte:fragment>
+							<div class="w-full">
+								<div class="flex flex-row items-center gap-2 text-sm font-semibold">
+									Runnable Inputs
 
-									<InputsSpecsEditor
-										id={component.id}
-										shouldCapitalize={false}
-										displayType
-										bind:inputSpecs={componentSettings.item.data.componentInput.fields}
-										userInputEnabled={component.type === 'formcomponent' ||
-											component.type === 'formbuttoncomponent'}
-									/>
-								</PanelSection>
+									<Tooltip wrapperClass="flex">
+										The runnable inputs are inferred from the inputs of the flow or script
+										parameters this component is attached to.
+									</Tooltip>
+								</div>
+
+								<InputsSpecsEditor
+									id={component.id}
+									shouldCapitalize={false}
+									displayType
+									bind:inputSpecs={componentSettings.item.data.componentInput.fields}
+									userInputEnabled={component.type === 'formcomponent' ||
+										component.type === 'formbuttoncomponent'}
+								/>
 							</div>
 						{/if}
 					{/if}
@@ -234,7 +298,7 @@
 			</PanelSection>
 		{/if}
 		{#if Object.values(initialConfiguration).length > 0}
-			<PanelSection title={`Configuration (${Object.values(initialConfiguration).length})`}>
+			<PanelSection title="Configuration">
 				<InputsSpecsEditor
 					id={component.id}
 					inputSpecsConfiguration={initialConfiguration}