+++ frontend/src/lib/components/apps/editor/settingsPanel/script/shared/ScriptTriggers.svelte
@@ -1,12 +1,13 @@
 <script lang="ts">
 	import type { InputConnection } from '$lib/components/apps/inputType'
 	import Alert from '$lib/components/common/alert/Alert.svelte'
-	import { classNames } from '$lib/utils'
+	import { classNames, itemsExists } from '$lib/utils'
 	import { Plus, X } from 'lucide-svelte'
 	import { Button } from '$lib/components/common'
 	import { getContext } from 'svelte'
 	import type { AppViewerContext, InlineScript } from '$lib/components/apps/types'
 	import Tooltip from '$lib/components/Tooltip.svelte'
+	import { deepEqual } from 'fast-equals'
 
 	export let triggerEvents: string[] = []
 	export let inlineScript: InlineScript | undefined = undefined
@@ -16,7 +17,7 @@
 
 	$: changeEvents = isFrontend
 		? inlineScript?.refreshOn
-			? inlineScript.refreshOn.map((x) => `${x.id} - ${x.key}`)
+			? inlineScript.refreshOn.map((x) => `${x.id}.${x.key}`)
 			: []
 		: dependencies
 
@@ -29,7 +30,7 @@
 		indigo: 'text-indigo-800 border-indigo-600 bg-indigo-100',
 		blue: 'text-blue-800 border-blue-600 bg-blue-100'
 	}
-	const { connectingInput, app } = getContext<AppViewerContext>('AppViewerContext')
+	const { connectingInput, app, stateId } = getContext<AppViewerContext>('AppViewerContext')
 
 	function applyConnection(connection: InputConnection) {
 		const refresh = {
@@ -47,37 +48,14 @@
 
 		if (!inlineScript.refreshOn) {
 			inlineScript.refreshOn = [refresh]
-		} else {
-			inlineScript.refreshOn.push(refresh)
+		} else if (!itemsExists(inlineScript.refreshOn, refresh)) {
+			inlineScript.refreshOn = [...inlineScript.refreshOn, refresh]
 		}
-
-		inlineScript = JSON.parse(JSON.stringify(inlineScript))
+		inlineScript = inlineScript
 		$app = $app
 	}
 </script>
 
-{#if isFrontend && shoudlDisplayChangeEvents}
-	<div class="flex mb-4">
-		<Button
-			size="xs2"
-			color="dark"
-			on:click={() => {
-				$connectingInput = {
-					opened: true,
-					input: undefined,
-					hoveredComponent: undefined,
-					onConnect: applyConnection
-				}
-			}}
-		>
-			<div class="flex flex-row gap-1 items-center">
-				<Plus size={14} />
-
-				Add dependency
-			</div>
-		</Button>
-	</div>
-{/if}
 {#if hasNoTriggers}
 	<Alert type="warning" title="No triggers" size="xs">
 		This script has no triggers. It will never run.
@@ -87,12 +65,12 @@
 		<div class="text-xs font-semibold text-secondary mb-1">Events</div>
 		<div class="flex flex-row gap-2 flex-wrap">
 			{#each triggerEvents as triggerEvent}
-				<span class={classNames(badgeClass, colors['green'])}>{triggerEvent}</span>
+				<span class={classNames(badgeClass)}>{triggerEvent}</span>
 			{/each}
 		</div>
 	{/if}
 	{#if changeEvents.length > 0 && shoudlDisplayChangeEvents}
-		<div class="text-xs font-semibold text-secondary mb-1 mt-2">Change on value</div>
+		<div class="text-xs font-semibold text-secondary mb-1 mt-2">Values watched</div>
 		<div class="flex flex-row gap-2 flex-wrap">
 			{#each changeEvents as changeEvent}
 				<span class={classNames(badgeClass, colors['blue'])}>
@@ -110,9 +88,19 @@
 							on:click={() => {
 								if (inlineScript?.refreshOn) {
 									inlineScript.refreshOn = inlineScript.refreshOn.filter(
-										(x) => `${x.id} - ${x.key}` !== changeEvent
+										(x) => `${x.id}.${x.key}` !== changeEvent
 									)
-									inlineScript = JSON.parse(JSON.stringify(inlineScript))
+									const ch = changeEvent.split('.')
+									const suggestion = {
+										id: ch[0],
+										key: ch[1]
+									}
+									if (!itemsExists(inlineScript.suggestedRefreshOn, suggestion)) {
+										inlineScript.suggestedRefreshOn = [
+											...(inlineScript.suggestedRefreshOn ?? []),
+											suggestion
+										]
+									}
 								}
 							}}
 						>
@@ -121,6 +109,54 @@
 					{/if}
 				</span>
 			{/each}
+		</div>
+	{/if}
+{/if}
+{#if isFrontend && shoudlDisplayChangeEvents}
+	<div class="flex my-4">
+		<Button
+			size="xs2"
+			color="dark"
+			on:click={() => {
+				$connectingInput = {
+					opened: true,
+					input: undefined,
+					hoveredComponent: undefined,
+					onConnect: applyConnection
+				}
+			}}
+		>
+			<div class="flex flex-row gap-1 items-center">
+				<Plus size={14} />
+				Add dependency
+			</div>
+		</Button>
+	</div>
+	{#if (inlineScript?.suggestedRefreshOn ?? []).length > 0}
+		<div class="gap-1 flex flex-wrap mb-2"
+			><span class="text-secondary text-sm">Quick add:</span>
+			{#key $stateId}
+				{#each inlineScript?.suggestedRefreshOn ?? [] as suggestion}
+					<button
+						class={classNames(
+							'p-0.5 rounded-md hover:bg-blue-400 cursor-pointer !text-2xs text-secondary',
+							badgeClass
+						)}
+						on:click={() => {
+							if (inlineScript) {
+								if (!itemsExists(inlineScript.refreshOn, suggestion)) {
+									inlineScript.refreshOn = [...(inlineScript.refreshOn ?? []), suggestion]
+									inlineScript.suggestedRefreshOn = inlineScript.suggestedRefreshOn?.filter(
+										(x) => !deepEqual(x, suggestion)
+									)
+								}
+							}
+						}}
+					>
+						+{suggestion.id}.{suggestion.key}
+					</button>
+				{/each}
+			{/key}
 		</div>
 	{/if}
 {/if}