+++ frontend/src/lib/components/apps/types.ts
@@ -14,6 +14,7 @@ import type {
 	ConnectedAppInput,
 	ConnectedInput,
 	EvalAppInput,
+	EvalV2AppInput,
 	InputConnection,
 	ResultAppInput,
 	RowAppInput,
@@ -34,7 +35,6 @@ export type Aligned = {
 }
 
 export interface GeneralAppInput {
-	onlyStatic?: boolean
 	tooltip?: string
 	placeholder?: string
 	customTitle?: string
@@ -55,6 +55,7 @@ export type Configuration =
 	| UserAppInput
 	| RowAppInput
 	| EvalAppInput
+	| EvalV2AppInput
 	| UploadAppInput
 	| ResultAppInput
 
@@ -106,6 +107,7 @@ export type InlineScript = {
 	path?: string
 	schema?: Schema
 	refreshOn?: { id: string; key: string }[]
+	suggestedRefreshOn?: { id: string; key: string }[]
 }
 
 export type AppCssItemName = 'viewer' | 'grid' | AppComponent['type']